Feature/coverage (#4)
* Improved readability & reliability of test_test_functions.sh

Tests that check test log output incur in a race condition: the log writer subprocess
may have not finished its job at the time of the check. Moreover, the code to filter the
interesting lines in the log makes it obscure and fragile.

* Re-source all scripts when invoking subshells

This serves two purposes:
* The tool bashcov correctly keeps track of coverage
* Stack traces show lines in source files

Several alternatives tried, the least dangerous seems to exhibit a bug in bash.

* Keep call stack accross subshells

* Fix stacking of stacks

* Remove calls to subshell from the call stack

* Clean logged output of test_log.sh

* Added safety check for infinite loops on invalid reentering

* Added codecov Makefile target

* Make reentrant subshell behaviour configurable

* Improved variable substitution in build target

* Fix config variable REENTER

* Added codecov_clean Makefile target

* Fix codecov target: report coverage of test.sh

* Run coverage and upload artifacts in ci workflow

* Install requirements for coverage in workflow agent

* Generate and publsh coverage badge in ci workflow

* eepa

* aaa

* fsdfsd

* sanitize

* sanitize

* sanitize

* badge color

* badge color

* Divide jobs

* Fix job definition

* Fix job definition

* Fix job definition

* Test workflow

+++ test/test_FAIL_FAST.sh
@@ -1,5 +1,3 @@
-source "$(dirname "$(readlink -f "$0")")"/../test.sh
-
 test_ok() {
   touch "$TEST_SCRIPT_DIR"/.test_ok
   TEST_OK=1
@@ -13,19 +11,22 @@ test_fail() {
   true
 }
 
+[ "$REENTRANT" != 1 ] || return 0
+source "$(dirname "$(readlink -f "$0")")"/../test.sh
+
 set_test_name "Any command that fails in the body of a test function should make the test to fail"
 rm -f "$TEST_SCRIPT_DIR"/.test_ok "$TEST_SCRIPT_DIR"/.test_fail
-! CURRENT_TEST_NAME= bash -c "run_tests test_fail"
+! CURRENT_TEST_NAME= subshell "run_tests test_fail"
 [ -f "$TEST_SCRIPT_DIR"/.test_fail ]
 
 set_test_name "FAIL_FAST should interrupt the script at the first test failure"
 rm -f "$TEST_SCRIPT_DIR"/.test_ok "$TEST_SCRIPT_DIR"/.test_fail
-! CURRENT_TEST_NAME= bash -c "run_tests test_fail test_ok 3>/dev/null"
+! CURRENT_TEST_NAME= subshell "run_tests test_fail test_ok 3>/dev/null"
 ! [ -f "$TEST_SCRIPT_DIR"/.test_ok ]
 [ -f "$TEST_SCRIPT_DIR"/.test_fail ]
 
 set_test_name "not FAIL_FAST should run all tests but signal failure at the end"
 rm -f "$TEST_SCRIPT_DIR"/.test_ok "$TEST_SCRIPT_DIR"/.test_fail
-! FAIL_FAST=0 CURRENT_TEST_NAME= bash -c "run_tests test_fail test_ok"
+! FAIL_FAST=0 CURRENT_TEST_NAME= subshell "run_tests test_fail test_ok"
 [ -f "$TEST_SCRIPT_DIR"/.test_ok ]
 [ -f "$TEST_SCRIPT_DIR"/.test_fail ]
